requirements.txt
astroid==2.0.4
autopep8==1.4.3
backoff==1.6.0
beautifulsoup4==4.6.3
biopython==1.72
certifi==2018.10.15
chardet==3.0.4
Click==7.0
dash==0.30.0
dash-core-components==0.38.0
dash-html-components==0.13.2
dash-renderer==0.15.0
dash-table==3.1.6
decorator==4.3.0
Flask==1.0.2
Flask-Compress==1.4.0
idna==2.7
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
langcodes==1.4.1
lazy-object-proxy==1.3.1
lxml==4.2.5
marisa-trie==0.7.5
MarkupSafe==1.1.0
mccabe==0.6.1
msgpack==0.5.6
nbformat==4.4.0
networkx==2.2
nltk==3.3
numpy==1.15.2
pandas==0.23.4
plotly==3.4.1
pycodestyle==2.4.0
pylint==2.1.1
python-dateutil==2.7.5
pytz==2018.7
regex==2018.2.21
requests==2.20.1
retrying==1.3.3
six==1.11.0
snowballstemmer==1.2.1
textblob==0.15.1
traitlets==4.3.2
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wordfreq==2.2.0
wrapt==1.10.11
